+++ docs/CHANGELOG.md
@@ -3,6 +3,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com)
 
 ## [Unreleased] 
+### Added
+- Support for m3 to wavefront obj export
+
 ## [0.1.3]
 ### Added
 - Convert command 'convert-file'. Allows to directly convert exported or otherwise acquired binary data from Wildstar.